+++ app/src/main/java/com/example/myapp/pages/content/Home.kt
@@ -1,9 +1,10 @@
-package com.example.myapp.pages
+package com.example.myapp.pages.content
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun OverView() {
-    Text("Overview Page")
+fun HomeView() {
+    Text("Home Page")
+
 }
