+++ app/src/main/java/com/example/myapp/pages/content/My.kt
@@ -1,10 +1,9 @@
-package com.example.myapp.pages
+package com.example.myapp.pages.content
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun HomeView() {
-    Text("Home Page")
-
+fun MyView() {
+    Text("My Page")
 }
