+++ app/src/main/java/com/example/myapp/pages/content/Overview.kt
@@ -1,9 +1,9 @@
-package com.example.myapp.pages
+package com.example.myapp.pages.content
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 
 @Composable
-fun MyView() {
-    Text("My Page")
+fun OverView() {
+    Text("Overview Page")
 }
