+++ app/src/main/java/com/example/myapp/pages/content/Scene.kt
@@ -1,6 +1,5 @@
-package com.example.myapp.pages
+package com.example.myapp.pages.content
 
-import android.transition.Scene
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
